Extract BloomTemplate and minor UI tweaks
