+++ bloom/src/main/java/com/fjoglar/composechallenge/bloom/MainActivity.kt
@@ -3,12 +3,9 @@ package com.fjoglar.composechallenge.bloom
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.material.MaterialTheme
-import androidx.compose.ui.graphics.Color
 import androidx.core.view.WindowCompat
+import com.fjoglar.composechallenge.bloom.ui.components.BloomTemplate
 import com.fjoglar.composechallenge.bloom.ui.screens.home.HomeScreen
-import com.fjoglar.composechallenge.bloom.ui.theme.BloomTheme
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -17,14 +14,7 @@ class MainActivity : ComponentActivity() {
         WindowCompat.setDecorFitsSystemWindows(window, false)
 
         setContent {
-            val systemUiController = rememberSystemUiController()
-
-            systemUiController.setStatusBarColor(
-                color = Color.Transparent,
-                darkIcons = MaterialTheme.colors.isLight
-            )
-
-            BloomTheme {
+            BloomTemplate {
                 HomeScreen()
             }
         }
